Add new createKeyframes/createAnimations
createKeyframes now takes a single keyframe object with percents as keys and returns the animation name whose rule was injected into the page.
createAnimations takes an object with animation names as keys and keyframe objects as values and returns an object mapping animation names in the source object to the dest names injected into the page.
Update to react@0.14.4, es2015 more things.